+++ core/autoscaling_test.go
@@ -1398,6 +1398,141 @@ func TestNeedReplaceOnDemandInstances(t *testing.T) {
 	}
 }
 
+func TestDetachAndTerminateOnDemandInstance(t *testing.T) {
+	tests := []struct {
+		name         string
+		instancesASG instances
+		regionASG    *region
+		instanceID   *string
+		expected     error
+	}{
+		{name: "no err during detach nor terminate",
+			instancesASG: makeInstancesWithCatalog(
+				map[string]*instance{
+					"1": {
+						Instance: &ec2.Instance{
+							InstanceId: aws.String("1"),
+							State: &ec2.InstanceState{
+								Name: aws.String(ec2.InstanceStateNameRunning),
+							},
+						},
+						region: &region{
+							services: connections{
+								ec2: mockEC2{tierr: nil},
+							},
+						},
+					},
+				},
+			),
+			regionASG: &region{
+				name: "regionTest",
+				services: connections{
+					autoScaling: mockASG{dierr: nil},
+				},
+				conf: &Config{},
+			},
+			instanceID: aws.String("1"),
+			expected:   nil,
+		},
+		{name: "err during detach not during terminate",
+			instancesASG: makeInstancesWithCatalog(
+				map[string]*instance{
+					"1": {
+						Instance: &ec2.Instance{
+							InstanceId: aws.String("1"),
+							State: &ec2.InstanceState{
+								Name: aws.String(ec2.InstanceStateNameRunning),
+							},
+						},
+						region: &region{
+							services: connections{
+								ec2: mockEC2{tierr: nil},
+							},
+						},
+					},
+				},
+			),
+			regionASG: &region{
+				name: "regionTest",
+				services: connections{
+					autoScaling: mockASG{dierr: errors.New("detach")},
+				},
+				conf: &Config{},
+			},
+			instanceID: aws.String("1"),
+			expected:   errors.New("detach"),
+		},
+		{name: "no err during detach but error during terminate",
+			instancesASG: makeInstancesWithCatalog(
+				map[string]*instance{
+					"1": {
+						Instance: &ec2.Instance{
+							InstanceId: aws.String("1"),
+							State: &ec2.InstanceState{
+								Name: aws.String(ec2.InstanceStateNameRunning),
+							},
+						},
+						region: &region{
+							services: connections{
+								ec2: mockEC2{tierr: errors.New("terminate")},
+							},
+						},
+					},
+				},
+			),
+			regionASG: &region{
+				name: "regionTest",
+				services: connections{
+					autoScaling: mockASG{dierr: nil},
+				},
+				conf: &Config{},
+			},
+			instanceID: aws.String("1"),
+			expected:   errors.New("terminate"),
+		},
+		{name: "errors during detach and terminate",
+			instancesASG: makeInstancesWithCatalog(
+				map[string]*instance{
+					"1": {
+						Instance: &ec2.Instance{
+							InstanceId: aws.String("1"),
+							State: &ec2.InstanceState{
+								Name: aws.String(ec2.InstanceStateNameRunning),
+							},
+						},
+						region: &region{
+							services: connections{
+								ec2: mockEC2{tierr: errors.New("terminate")},
+							},
+						},
+					},
+				},
+			),
+			regionASG: &region{
+				name: "regionTest",
+				services: connections{
+					autoScaling: mockASG{dierr: errors.New("detach")},
+				},
+				conf: &Config{},
+			},
+			instanceID: aws.String("1"),
+			expected:   errors.New("detach"),
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			a := autoScalingGroup{
+				name:      "testASG",
+				region:    tt.regionASG,
+				instances: tt.instancesASG,
+			}
+			err := a.detachAndTerminateOnDemandInstance(tt.instanceID)
+			CheckErrors(t, err, tt.expected)
+		})
+	}
+}
+
 func TestTerminateInstanceInAutoScalingGroup(t *testing.T) {
 	tests := []struct {
 		name         string
@@ -2262,8 +2397,10 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 										tierr: nil,
 									},
 									autoScaling: &mockASG{
-										aio:   nil,
-										aierr: nil,
+										aio:       nil,
+										aierr:     nil,
+										tiiasgo:   nil,
+										tiiasgerr: nil,
 									},
 								},
 							},
@@ -2293,6 +2430,8 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 						autoScaling: &mockASG{
 							uasgo:     nil,
 							uasgerr:   nil,
+							dio:       nil,
+							dierr:     nil,
 							tiiasgo:   nil,
 							tiiasgerr: nil,
 						},
@@ -2393,6 +2532,8 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 						autoScaling: &mockASG{
 							uasgo:     nil,
 							uasgerr:   nil,
+							dio:       nil,
+							dierr:     nil,
 							tiiasgo:   nil,
 							tiiasgerr: nil,
 						},
@@ -2441,6 +2582,8 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 						autoScaling: &mockASG{
 							uasgo:     nil,
 							uasgerr:   nil,
+							dio:       nil,
+							dierr:     nil,
 							tiiasgo:   nil,
 							tiiasgerr: nil,
 						},
@@ -2494,6 +2637,8 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 						autoScaling: &mockASG{
 							uasgo:     nil,
 							uasgerr:   nil,
+							dio:       nil,
+							dierr:     nil,
 							tiiasgo:   nil,
 							tiiasgerr: nil,
 						},
@@ -2525,6 +2670,11 @@ func TestReplaceOnDemandInstanceWithSpot(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			returned := tt.asg.replaceOnDemandInstanceWithSpot(tt.spotID)
+			CheckErrors(t, returned, tt.expected)
+		})
+		t.Run(tt.name+"-detach-method", func(t *testing.T) {
+			tt.asg.terminationMethod = "detach"
 			returned := tt.asg.replaceOnDemandInstanceWithSpot(tt.spotID)
 			CheckErrors(t, returned, tt.expected)
 		})