+++ terraform/autospotting/lambda/main.tf
@@ -15,6 +15,7 @@ resource "aws_lambda_function" "autospotting" {
     variables = {
       ALLOWED_INSTANCE_TYPES       = "${var.autospotting_allowed_instance_types}"
       DISALLOWED_INSTANCE_TYPES    = "${var.autospotting_disallowed_instance_types}"
+      INSTANCE_TERMINATION_METHOD  = "${var.autospotting_instance_termination_method}"
       MIN_ON_DEMAND_NUMBER         = "${var.autospotting_min_on_demand_number}"
       MIN_ON_DEMAND_PERCENTAGE     = "${var.autospotting_min_on_demand_percentage}"
       ON_DEMAND_PRICE_MULTIPLIER   = "${var.autospotting_on_demand_price_multiplier}"
@@ -45,6 +46,7 @@ resource "aws_lambda_function" "autospotting_from_s3" {
     variables = {
       ALLOWED_INSTANCE_TYPES       = "${var.autospotting_allowed_instance_types}"
       DISALLOWED_INSTANCE_TYPES    = "${var.autospotting_disallowed_instance_types}"
+      INSTANCE_TERMINATION_METHOD  = "${var.autospotting_instance_termination_method}"
       MIN_ON_DEMAND_NUMBER         = "${var.autospotting_min_on_demand_number}"
       MIN_ON_DEMAND_PERCENTAGE     = "${var.autospotting_min_on_demand_percentage}"
       ON_DEMAND_PRICE_MULTIPLIER   = "${var.autospotting_on_demand_price_multiplier}"